correção ex008.py

+++ ex008.py
@@ -1,4 +1,4 @@
-metro = int(input('Digite um valor em metros: '))
+metro = float(input('Digite um valor em metros: '))
 cm = metro * 100
 mm = metro * 1000
 print('{} metro(s) convertido em centímetros é igual a {} cm.'.format(metro, cm))
